Refactor monetary fields and price calculations to use `BigDecimal` for precision.
- Updated DTOs (`OrderDTO`, `PreOrderResponseDTO`, `CartResponseDTO`, `BucketDTO`, `OrderItemDTO`, `OrderResponseDTO`) and Entities (`OrderItem`, `PreOrder`, `Order`) to replace `double` with `BigDecimal` for all price, discount, and amount fields.
- Refactored `Constants.calculateUnitPrice` and price calculation logic in `PreOrderServiceImpl`, `OrderServiceImpl`, and `BucketServiceImpl` to use `BigDecimal` arithmetic and `RoundingMode.HALF_UP`.
- Updated `PreOrderController` to validate `amountToPay` using `BigDecimal` comparisons.
- Modified `Stock` and `MedicineDTO` to use `LocalDate` for `expiryDate` instead of `String`.
- Added JPA column mappings for `Vendor` and `Stock` entities to align with database naming conventions.
- Improved bucket sorting logic in `BucketServiceImpl` to use `Comparator.comparing` with `BigDecimal`.

+++ src/main/java/com/trego/dao/entity/Order.java
@@ -27,10 +27,10 @@ public class Order {
     private Vendor vendor;
 
     @Column(nullable = false)
-    private double totalAmount;
+    private BigDecimal totalAmount;
 
     @Column(nullable = false)
-    private double discount;
+    private BigDecimal discount;
 
     @Column(nullable = false)
     private String address;
