+++ src/main/java/com/trego/dto/response/OrderResponseDTO.java
@@ -6,6 +6,7 @@ import com.trego.dao.entity.Order;
 import com.trego.dto.AddressDTO;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
@@ -14,9 +15,9 @@ public class OrderResponseDTO {
     private Long userId;
     private Long orderId;
     private String razorpayOrderId;
-    private double totalCartValue;
-    private double amountToPay;
-    private double discount;
+    private BigDecimal totalCartValue;
+    private BigDecimal amountToPay;
+    private BigDecimal discount;
     private String mobileNo;
     private AddressDTO address;
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd HH:mm:ss")
